+++ config/constants.js
@@ -13,19 +13,6 @@ const editButton = document.querySelector(".edit-confirmation-btn");
 const buttonRegistration = document.querySelector(".button-registration");
 const buttonLogin = document.querySelector(".button-login");
 
-// const routes = {
-//     home: {
-//         template: "/templates/index.html",
-//         title: "Home",
-//         controller: () => { },
-//     },
-//     about: {
-//         template: "/templates/about.html",
-//         title: "About Us",
-//         controller: () => { },
-//     },
-// };
-
 const phone = "+980999999998";
 
 export {
@@ -39,6 +26,5 @@ export {
     buttonLogin,
     modalLogin,
     modalRegistration,
-    // routes,
     phone,
 }
