+++ index.html
@@ -5,13 +5,13 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Document</title>
-    <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="./styles/style.css" />
   </head>
   <body>
     <div class="wraper">
       <header class="header">
-        <button class="button-registration">Registration</button>
-        <button class="button-login">Login</button>
+        <a href="#registration" class="button-registration">Registration</a>
+        <a href="#login" class="button-login">Login</a>
         <a href="#home">Home</a>
         <a href="#about">About</a>
         <a href="#contact">Contact</a>
